Closes NTWK-348: Create DNS name on VIP creation
This patch places a notification on the MQ with a paylod
containing the tuple {vip_id, vip_IPaddress, vip_hostname,
vip_tenant_id, vip_DNS_zone_name} as required by the DNS
controller to create an APTR record for the newly created
VIP. The hostname for the VIP is generated using the name
field of the VIP, appended with a 5 character random string.
The class of service of a tenant is obtained by making a
call to keystone with the tenant-get <tenant_id> API.

A new config parameter in neutron.conf, vpc_dns_zone_names
is used to look up the DNS zone name of a specific class of
service.

Change-Id: I9fca10d3254acdd4e97ba3fce195a4edc23514a2